pyproject.toml
[tool.poetry]
name = "myfinance"
version = "1.0.0"
description = "Aplicação para controle financeiro"
authors = ["Fernando Antunes <ferantunesdev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
asgiref = "3.7.2"
certifi = "2023.7.22"
cffi = "1.15.1"
chardet = "5.2.0"
charset-normalizer = "3.2.0"
cryptography = "41.0.3"
defusedxml = "0.7.1"
django = "4.2.5"
django-upload-validator = "1.1.6"
djangorestframework = "3.14.0"
djangorestframework-simplejwt = "5.3.0"
et-xmlfile = "1.1.0"
hvac = "1.2.1"
idna = "3.4"
mysqlclient = "2.2.0"
oauthlib = "3.2.2"
openpyxl = "3.1.2"
pillow = "10.0.1"
psycopg2 = "2.9.7"
pycparser = "2.21"
pyhcl = "0.4.5"
pyjwt = "2.8.0"
python-dateutil = "2.8.2"
python-dotenv = "1.0.0"
python-magic = "0.4.27"
python3-openid = "3.2.0"
pytz = "2023.3.post1"
requests = "2.31.0"
requests-oauthlib = "1.3.1"
six = "1.16.0"
social-auth-app-django = "5.3.0"
social-auth-core = "4.4.2"
sqlparse = "0.4.4"
urllib3 = "2.0.4"
gunicorn = "^22.0.0"


[tool.poetry.group.dev.dependencies]
blue = "^0.9.1"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.blue]
line-lenght = 100
